Starting to port pymemdyn to gromacs greater than or equal to five The logic remains almost the same, there's only need for a change in the command list in the generate_command function for the Wrapper class of gromacs.py, the list which used to have just one element now has two, including gmx or gmx_mpi, or gmx_d or whichever is the new prepended command, this is now defined in the settings file.
A problem which surfaces is that genbox does not exist anymore in
more recent versions of gromacs, so gmx solvation must be used.

Unfortunately, gmx solvation has a redmine bug, see:

<https://redmine.gromacs.org/issues/1887>

So, probably some releases of gromacs 2016 will not be compatible at all.

Yet another problem seems to be that the new gmx ignores the usual behavior of
c for argument variables, and something funky is happening with stderr and the
whole pipeline is going through ignoring the error raising parts of
the gromacs.py module.