Still at eq.mdp One step forward, sort-of. New versions of gromacs only accept one value in tau_p. This seems to be a bug, but, since for our case the value is the same, then it's not a problem and it has been commented-out.
https://mailman-1.sys.kth.se/pipermail/gromacs.org_gmx-users/2016-February/103458.html

Now, after solving this, then a bigger problem is aparent.
That bigger problem is that, perhaps the new *gmx genion* is killing
the index groups in the topology.
The index groups wation protlig and membr are gone at some
point from the topology and then the equilibration breaks because it can't
find them.